feat(mcp): host the MCP tool set on drsg serve at /mcp (ROADMAP §10)
Implements the design settled in ROADMAP.md §10: several agent hosts
sharing one memory needs the MCP transport hosted by `drsg serve`, not a
`--connect` proxy in `drsg-mcp` — the tools don't map cleanly onto /rpc
(traverse has no method, write_nodes/write_edges lose batch atomicity as N
round-trips, digest needs composing from two calls).

dr-strange-mcp gains a lib target (DrStrange and the transport-independent
`*_logic` functions behind its 15 tools) alongside its existing stdio
binary. dr-strange-web depends on it and mounts the same DrStrange service
at POST /mcp over rmcp's Streamable HTTP transport, gated by the same
DRSG_TOKEN bearer auth /rpc already uses (Access::Write — the v1
single-token model doesn't distinguish finer, matching how /cypher is
already gated). One DrStrange instance per MCP session, each holding the
same Arc<Database> the rest of the server shares, so a write from one
session is immediately visible to another — proven by an end-to-end test
driving two independent rmcp client sessions against a live server, not
just that the request JSON is shaped right.

The new dependency is taken with `default-features = false`, for the reason
[workspace.dependencies] already states for core: the storage backend is
the depending crate's to forward. Without it, dr-strange-mcp's default
(`native-backend`) reaches core through this edge even in a
`--no-default-features --features redb-backend` build, and core's
`cfg(all(redb-backend, not(native-backend)))` would then select the native
engine with nothing to say so.

drsg-mcp itself is unchanged: still embedded, still stdio-only, still the
right answer for one agent. A host that wants the shared server points its
own MCP client straight at the URL — no client mode needed in drsg-mcp for
that.

One asymmetry is documented and pinned rather than papered over: rmcp's
Streamable HTTP transport validates the inbound Host header against a
loopback-only list to blunt DNS rebinding, so `/mcp` answers 403 on a
hostname where `/rpc` on the same server answers normally. The guard is
worth keeping — a tokenless server trusts its own same-origin UI, which is
what rebinding impersonates — and same-machine hosts, the §10 case, are
unaffected. Making it configurable would be a ServeOptions decision, so
this change only states the limit and tests it.

docs/{en,zh}/src/mcp.md gain a "Sharing across agents" section, extending
1.4.2's "should not run against a database drsg serve currently holds open"
language with what to do instead. Closes the second half of #1, whose other
half (the cross-process lock) shipped in 1.4.2.
