+++ docs/en/src/mcp.md
@@ -69,10 +69,65 @@ not two at once. This is enforced, not advisory: the second open fails with a
 clear error rather than corrupting the database.
 
 That matters for agent hosts, because each one spawns its own MCP server
-subprocess. Two editors open on the same project means two `drsg-mcp` processes,
-and the second will refuse to start. Until `drsg-mcp` can attach to a running
-server, point each host at its own database, or drive one `drsg serve` through
-the SDKs or the RPC API for anything that must be shared.
+subprocess. Two editors open on the same project means two `drsg-mcp`
+processes, and the second will refuse to start. For that case — several
+agents that need to share one memory — see the next section.
+
+## Sharing across agents: `/mcp` on `drsg serve`
+
+`drsg-mcp` embeds by design: point a host at a path and it works, with
+nothing to run and nothing to configure. That's the right answer for one
+agent, but it's also *why* two hosts can't share a database directly — each
+one opens the file itself.
+
+`drsg serve` (arch/08) already solves the "one writer, several clients"
+problem for the JSON-RPC API and the dashboard. It exposes the identical
+tool set at `POST /mcp`, over MCP's Streamable HTTP transport, so several
+agent hosts can point at the same server instead of each embedding their own
+copy of the database:
+
+```json
+{
+  "mcpServers": {
+    "dr-strange": {
+      "url": "http://127.0.0.1:7700/mcp",
+      "headers": { "Authorization": "Bearer <DRSG_TOKEN>" }
+    }
+  }
+}
+```
+
+(Check your host's docs for its exact remote-MCP-server config shape — the
+`url`/`headers` fields above are illustrative, not universal.)
+
+`/mcp` is gated exactly like `/rpc`: with `DRSG_TOKEN` unset, only the
+same-origin browser UI is trusted, and **every programmatic client is
+refused, reads included** — a zero-config desktop install shouldn't quietly
+expose an open API on localhost. Set `DRSG_TOKEN` on the server and pass it
+as a bearer token to reach `/mcp` from anywhere else, including another
+agent host.
+
+One limit worth knowing before putting this behind a hostname: the MCP
+transport validates the inbound `Host` header against a loopback-only list
+(`localhost`, `127.0.0.1`, `::1`) to blunt DNS-rebinding attacks on locally
+running servers, and answers **403** to anything else. The check earns its
+keep here — a tokenless server trusts its own same-origin UI, which is what
+rebinding sets out to impersonate — but it does mean `/mcp` answers at
+`http://127.0.0.1:7700/mcp` and not at `https://memory.example.com/mcp`,
+where `/rpc` on the same server would. Agent hosts on the same machine, the
+case ROADMAP §10 is about, are unaffected.
+
+Every session gets its own `DrStrange` instance, but they all share the one
+`Database` the server opened — a write from one session is visible to every
+other, immediately, the same way two browser tabs against `/rpc` already
+are. That's the whole point: `write_gate` inside the core serializes
+concurrent writers so this is safe, not just convenient.
+
+`digest`'s LLM calls still spend the server process's provider keys (never a
+tool parameter, remote or local); `write_nodes`/`write_edges` keep their
+per-call batch atomicity, since the tool code runs in the same process
+against the same `Database` either way — nothing here proxies to `/rpc` or
+reshapes what a tool does.
 
 ## The tools
