+++ docs/zh/src/mcp.md
@@ -62,9 +62,55 @@ $ drsg-mcp /path/to/graph.drsg
 建议：第二次打开会以明确的错误失败，而不是把数据库弄坏。
 
 这一点对智能体宿主尤为要紧，因为每个宿主都会各自派生自己的 MCP 服务子进程。同一个项目
-开着两个编辑器，就意味着两个 `drsg-mcp` 进程，而后者将拒绝启动。在 `drsg-mcp` 能够连接
-到一个正在运行的服务之前，请为每个宿主分别指定各自的数据库；凡需共享的场景，则通过
-SDK 或 RPC 接口去驱动同一个 `drsg serve`。
+开着两个编辑器，就意味着两个 `drsg-mcp` 进程，而后者将拒绝启动。若干智能体需要共享同一
+份记忆的场景，见下一节。
+
+## 跨智能体共享：`drsg serve` 上的 `/mcp`
+
+`drsg-mcp` 按设计内嵌运行：把宿主指向一个路径即可工作，无需运行任何附加服务，也无需
+配置。这对单个智能体是正确答案，但也正因如此，两个宿主无法直接共享一个数据库——因为
+各自都在打开这个文件。
+
+`drsg serve`（arch/08）已经为 JSON-RPC 接口和仪表盘解决了"一个写者、多个客户端"的问题。
+它在 `POST /mcp` 上暴露完全相同的工具集，走 MCP 的 Streamable HTTP 传输，让多个智能体
+宿主可以指向同一个服务，而不必各自内嵌一份数据库副本：
+
+```json
+{
+  "mcpServers": {
+    "dr-strange": {
+      "url": "http://127.0.0.1:7700/mcp",
+      "headers": { "Authorization": "Bearer <DRSG_TOKEN>" }
+    }
+  }
+}
+```
+
+（请查阅你所用宿主的文档以确认其远程 MCP 服务配置的确切格式——以上 `url`/`headers`
+字段仅为示例，并非通用格式。）
+
+`/mcp` 的鉴权方式与 `/rpc` 完全一致：未设置 `DRSG_TOKEN` 时，只有同源的浏览器 UI
+受信任，**任何程序化客户端都会被拒绝，读操作也不例外**——零配置的桌面安装不应在
+本地悄悄暴露一个开放的 API。请在服务端设置 `DRSG_TOKEN`，并将其作为 bearer token
+传递，以便从任何其他地方（包括另一个智能体宿主）访问 `/mcp`。
+
+把它放到某个域名后面之前，有一处限制值得先知道：MCP 传输层会校验请求的 `Host` 头，
+只放行回环地址（`localhost`、`127.0.0.1`、`::1`），以此削弱针对本机服务的 DNS
+重绑定攻击，其余一律返回 **403**。这道检查在这里是有价值的——未设 token 的服务端
+信任自己的同源 UI，而这正是重绑定要冒充的东西——但它也意味着 `/mcp` 只在
+`http://127.0.0.1:7700/mcp` 上应答，而不在 `https://memory.example.com/mcp` 上，
+尽管同一个服务端的 `/rpc` 可以。ROADMAP §10 所针对的场景——同一台机器上的多个
+智能体宿主——不受影响。
+
+每个会话都会得到自己的 `DrStrange` 实例，但它们共享服务端打开的同一个 `Database`——
+一个会话的写入会立即对其他所有会话可见，正如两个浏览器标签页对 `/rpc` 的访问那样。
+这正是关键所在：核心内部的 `write_gate` 会序列化并发写者，因此这是安全的，而不仅仅是
+方便。
+
+`digest` 的 LLM 调用仍然花费服务进程自身的提供方密钥（无论本地还是远程，都绝不来自
+工具参数）；`write_nodes`/`write_edges` 保留其逐次调用的批量原子性，因为无论哪种情况，
+工具代码都在同一进程内针对同一个 `Database` 运行——这里不会代理到 `/rpc`，也不会改变
+任何工具的行为。
 
 ## 工具
